Fix sonar indicated issues
