+++ example/docsify-glossary.js
@@ -5,8 +5,41 @@
  * (c) 2018-2022 Stijn Dejongh
  * Apache-2.0 license
  */
-(function() {
+this["@stijn-dejongh/docsify-glossary"] = this["@stijn-dejongh/docsify-glossary"] || {};
+
+this["@stijn-dejongh/docsify-glossary"].js = function(exports) {
     "use strict";
+    function replaceTerm(term, content, term_id) {
+        var link = " [$1](/_glossary?id=".concat(term_id, ")");
+        var re = new RegExp("\\s(".concat(term, ")\\s"), "ig");
+        var reFullStop = new RegExp("\\s(".concat(term, ")."), "ig");
+        var reComma = new RegExp("\\s(".concat(term, "),"), "ig");
+        return content.replace(reComma, link + ",").replace(re, link + " ").replace(reFullStop, link + ".");
+    }
+    function addLinks(content, terms, config) {
+        var textWithReplacements = content;
+        if (config.debug) {
+            console.log("Adding links for terminology: ".concat(terms));
+        }
+        for (var term in terms) {
+            textWithReplacements = replaceTerm(term, textWithReplacements, terms[term]);
+        }
+        return textWithReplacements;
+    }
+    function loadTerminology(text, configuration) {
+        var lines = text.split("\n");
+        var dictionary = {};
+        lines.forEach((function(line) {
+            if (line.trimStart().startsWith(configuration.terminologyHeading)) {
+                var term = line.replace(configuration.terminologyHeading, "").trim();
+                if (configuration.debug) {
+                    console.log("detected glossary term: ".concat(term));
+                }
+                dictionary[term] = term.toLowerCase().replace(" ", "-");
+            }
+        }));
+        return dictionary;
+    }
     function ownKeys(object, enumerableOnly) {
         var keys = Object.keys(object);
         if (Object.getOwnPropertySymbols) {
@@ -104,51 +137,20 @@
         var _configurationYaml$te, _configurationYaml$gl, _configurationYaml$de;
         return (new GlossaryConfigurationBuilder).withTermHeading((_configurationYaml$te = configurationYaml.terminologyHeading) !== null && _configurationYaml$te !== void 0 ? _configurationYaml$te : DEFAULT_TERM_HEADING).withGlossaryLocation((_configurationYaml$gl = configurationYaml.glossaryLocation) !== null && _configurationYaml$gl !== void 0 ? _configurationYaml$gl : DEFAULT_GLOSSARY_FILE_NAME).withDebugEnabled((_configurationYaml$de = configurationYaml.debug) !== null && _configurationYaml$de !== void 0 ? _configurationYaml$de : false).build();
     }
-    function glossifyConfig() {
-        return new GlossaryConfigurationBuilder;
+    function defaultGlossifyConfig() {
+        return (new GlossaryConfigurationBuilder).build();
     }
-    function replaceTerm(term, content, term_id) {
-        var link = " [$1](/_glossary?id=".concat(term_id, ")");
-        var re = new RegExp("\\s(".concat(term, ")\\s"), "ig");
-        var reFullStop = new RegExp("\\s(".concat(term, ")."), "ig");
-        var reComma = new RegExp("\\s(".concat(term, "),"), "ig");
-        return content.replace(reComma, link + ",").replace(re, link + " ").replace(reFullStop, link + ".");
-    }
-    function addLinks(content, terms, config) {
-        var textWithReplacements = content;
-        if (config.debug) {
-            console.log("Adding links for terminology: ".concat(terms));
-        }
-        for (var term in terms) {
-            textWithReplacements = replaceTerm(term, textWithReplacements, terms[term]);
-        }
-        return textWithReplacements;
-    }
-    function loadTerminology(text, configuration) {
-        var lines = text.split("\n");
-        var dictionary = {};
-        lines.forEach((function(line) {
-            if (line.trimStart().startsWith(configuration.terminologyHeading)) {
-                var term = line.replace(configuration.terminologyHeading, "").trim();
-                if (configuration.debug) {
-                    console.log("detected glossary term: ".concat(term));
-                }
-                dictionary[term] = term.toLowerCase().replace(" ", "-");
-            }
-        }));
-        return dictionary;
+    function injectTerminologyInContent(content, configuration, next) {
+        content = addLinks(content, window.$docsify.terms, configuration);
+        next(content);
     }
     function loadProperties() {
         if (window.$docsify !== undefined && window.$docsify.glossify !== undefined) {
             var configuredProperties = window.$docsify.glossify;
             return configFromYaml(configuredProperties);
         } else {
-            return glossifyConfig().build();
+            return defaultGlossifyConfig();
         }
-    }
-    function injectTerminologyInContent(content, configuration, next) {
-        content = addLinks(content, window.$docsify.terms, configuration);
-        next(content);
     }
     function install(hook, _vm) {
         var configuration = loadProperties();
@@ -175,4 +177,9 @@
         window.$docsify = {};
     }
     window.$docsify.plugins = (window.$docsify.plugins || []).concat(install);
-})();
+    exports.install = install;
+    Object.defineProperty(exports, "__esModule", {
+        value: true
+    });
+    return exports;
+}({});
