+++ sonar-project.properties
@@ -2,3 +2,4 @@ sonar.projectKey=stijn-dejongh_docsify-glossary
 sonar.organization=stijn-dejongh
 sonar.javascript.lcov.reportPaths=./coverage/lcov.info
 sonar.coverage.exclusions=**/*.spec.ts,**/*.spec.js,**/*.config.js,**/index.js
+sonar.exclusions=example/**
